MonitoredQuantity to aggregation.py
